+++ lib/openapi_first/header.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
+require_relative 'parameter'
+
 module OpenapiFirst
-  Header = Data.define(:name, :required?, :schema, :resolved_schema)
+  # A header of a response definition.
+  # @attr_reader [Parameter] parameter The header as a Parameter, which knows how to unpack a raw value.
+  Header = Data.define(:name, :required?, :schema, :parameter)
 end